% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distance.R
\name{cd}
\alias{cd}
\title{cd}
\usage{
cd(point1, point2)
}
\arguments{
\item{point1}{vector with numeric values}

\item{point2}{vector with numeric values}
}
\value{
float, the distance between point1 and point2 based on the cosine similarity
}
\description{
cd
}
\examples{

cd(c(0,0,0), c(1,0,0))

}
