% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_distances.R
\name{filter_distances}
\alias{filter_distances}
\title{filter_distances}
\usage{
filter_distances(point, data, threshold, metric = "euclidean")
}
\arguments{
\item{point}{Vector, values defining a single observation to compute distances for}

\item{data}{Dataframe, values of all observations to calculate distances from point}

\item{threshold}{float, maximum distance of observations to return indices for}

\item{metric}{Char, type of distance metric to use in distance calculations}
}
\value{
Vector, indices of the observations with distance less than `threshold` from `point`
}
\description{
filter_distances
}
\examples{
df <- data.frame(x=c(1, 1), y=c(1,2))
filter_distances(c(1,1), df, 0.9)

}
