% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_distances.R
\name{get_all_distances}
\alias{get_all_distances}
\title{Return distance/similarity metric for each row in a dataframe}
\usage{
get_all_distances(point, data, dist = "euclidean")
}
\arguments{
\item{point}{vector of length k to compare to the data frame}

\item{data}{data frame of size n by k to compare ref_vec to}

\item{dist}{string indicating type of distance metric}
}
\value{
numeric vector of length n
}
\description{
Created by Evan Yathon on 2019-02-08
}
\details{
Compares an input reference vector to all rows of an input data frame, calculating the specified distance/similarity metric for each row.
}
\examples{

}
