% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_closest.R
\name{get_closest}
\alias{get_closest}
\title{get_closest}
\usage{
get_closest(point, data, top_k, dist = "euclidean")
}
\arguments{
\item{point}{Vector, values defining a single observation to compute distances for}

\item{data}{Dataframe, values of all observations to calculate distances from point}

\item{top_k}{Integer, number of closest observations to return indices for}

\item{dist}{Char, type of distance metric to use in distance calculations}
}
\value{
Vector, indices of the closest k observations from data
}
\description{
get_closest
}
\examples{
df <- data.frame(x=c(1, 1), y=c(1,2))
get_closest(c(1,1), df, 2)

}
