% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distance.R
\name{get_distance}
\alias{get_distance}
\title{get_distance}
\usage{
get_distance(point1, point2, metric = "euclidean")
}
\arguments{
\item{point1}{vector with numeric values}

\item{point2}{vector with numeric values}

\item{metric}{default "euclidean", "cosine" will return cosine similiarity distance, and "manhattan" will return manhattan distance}
}
\value{
float, the distance between point1 and point2 based on the metric given
}
\description{
get_distance
}
\examples{

get_distance(c(0,0,0), c(1,0,0))

}
