% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distance.R
\name{man}
\alias{man}
\title{man}
\usage{
man(point1, point2)
}
\arguments{
\item{point1}{vector with numeric values}

\item{point2}{vector with numeric values}
}
\value{
float, the distance between point1 and point2 based on the manhattan
}
\description{
man
}
\examples{

man(c(0,0,0), c(1,0,0))

}
