% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_distances.R
\name{get_all_distances}
\alias{get_all_distances}
\title{Return distance/similarity metric for each row in a dataframe}
\usage{
get_all_distances(point, data, metric = "euclidean")
}
\arguments{
\item{point}{vector of length k to compare to the data frame}

\item{data}{data frame of size n by k to compare ref_vec to}

\item{metric}{string, default value "euclidean" for euclidean distance, "cosine" will return cosine similiarity distance, and "manhattan" will return manhattan distance}
}
\value{
numeric vector of length n
}
\description{
Created by Evan Yathon on 2019-02-08
}
\details{
Compares an input reference vector to all rows of an input data frame, calculating the specified distance/similarity metric for each row.
}
\examples{
df <- data.frame("A" = c(1,2,3), "B" = c(8,2,4))
ref_vec <- c(-2,4)
get_all_distances(ref_vec, df, metric = "euclidean")

}
